# swagger.yaml
swagger: '2.0'
info:
  title: RiftBattle API
  description: API for managing users, products, orders, Fortnite stats, and more.
  version: 1.0.0
basePath: /api
...
paths:
...
  /get_products:
    get:
      tags:
        - product
      summary: Get all products
      responses:
        '200':
          description: Successful response with products list
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /get_skins_info:
    get:
      tags:
        - fortnite
      summary: Get information about Fortnite skins
      parameters:
        - name: athenaID
          in: query
          type: string
          required: true
          description: Athena ID of the skin
      responses:
        '200':
          description: Successful response with skin info
          schema:
            $ref: '#/definitions/AthenaID'
        '400':
          description: Bad request (e.g., missing or invalid Athena ID)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "athenaID is required"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /seller/create_product:
    post:
      tags:
        - product
      summary: Create a new product (Seller only)
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                example: "Fortnite Account"
              price:
                type: number
                format: float
                example: 50.0
              athena_ids:
                type: array
                items:
                  type: string
                example: ["CID_123", "CID_456"]
              stats:
                type: object
                example: {"level": "100", "backpacks": "500"}
              active:
                type: boolean
                example: true
              description:
                type: string
                example: "A rare Fortnite account"
              discount:
                type: integer
                example: 10
              credentials:
                type: string
                example: "username:password"
      responses:
        '200':
          description: Product created successfully
        '400':
          description: Bad request (e.g., invalid parameters)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "invalid parameters"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
  /user/send_message:
    post:
      tags:
        - message
      summary: Send a message
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Hello, I have a question!"
              receiver_id:
                type: integer
                example: 2
      responses:
        '200':
          description: Message sent successfully
        '400':
          description: Bad request (e.g., invalid message or receiver ID)
          schema:
            type: object
            properties:
              error:
                type: string
                example: "invalid message"
        '500':
          description: Internal server error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "internal server error"
...
definitions:
  AthenaID:
    type: object
    properties:
      id:
        type: integer
        example: 1
      athena_id:
        type: string
        example: "CID_123"
      name:
        type: string
        example: "Skin Name"
      type:
        type: string
        example: "outfit"
      rarity:
        type: string
        example: "legendary"
      description:
        type: string
        example: "A rare skin"
...
  Product:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "Fortnite Account"
      price:
        type: number
        format: float
        example: 50.0
      credentials:
        type: string
        example: "username:password"
      seller_id:
        type: integer
        example: 2
      athena_ids:
        type: array
        items:
          type: string
        example: ["CID_123", "CID_456"]
      stats:
        type: object
        example: {"level": "100", "backpacks": "500"}
      description:
        type: string
        example: "A rare Fortnite account"
      date:
        type: integer
        format: int64
        example: 1690924800
      active:
        type: boolean
        example: true
      discount:
        type: integer
        example: 10
...